pyproject.toml
[project]
name = "policy-orchestrator"
version = "0.1.0"
description = "Control plane for multi-repo policy enforcement, agent contracts, secrets hygiene, and cross-repo coordination"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "click>=8.1",
]

[project.scripts]
devctl = "policy_orchestrator.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/policy_orchestrator"]

[tool.ruff]
target-version = "py312"
line-length = 100
